Add missing rocm_skinny_gemms kernel test to CI
